Backport the apr_shm_remove patch from mod_fcgid
Please have a look at the mod_fcgid mailing list for details.
The original patch can be found at:
    http://lists.freebsd.org/pipermail/freebsd-ports-bugs/2007-June/122731.html

The patch itself is only required for Apache 2.0.x servers as older versions of apxs don't support the apr_shm_remove function.

Status of this patch: This patch has not been tested due to lack of a test-system.